FROMGIT: arm64: dts: meson-g12a-sei510: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 11ad4dfa8fa9de442db0976880f6c482567a8f41
 https://git.kernel.org/pub/scm/linux/kernel/git/khilman/linux-amlogic.git v5.5/dt64)
Change-Id: I29da1dbb631b04090cce5649eea0ecf52bdc42aa