FROMGIT: arm64: dts: meson-sm1-sei610: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 2e09574d172e6da4ee251203660c5a56b0b4c068
 https://git.kernel.org/pub/scm/linux/kernel/git/khilman/linux-amlogic.git v5.5/dt64)
Change-Id: Idd515178a0e54f5d9f64928129dfde5a34f3d018