FROMGIT: arm64: dts: meson: g12: add a g12 layer
While the sm1 is very close to the g12a/b family, somethings apply
differently on the g12a/b and not the sm1. This introduce a new layer
of dtsi for part which apply to the g12a and g12b but not the sm1.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit c36a19042d2d4bab0fd75e5eb9c782b79bdbf5a8
 git://git.kernel.org/pub/scm/linux/kernel/git/khilman/linux-amlogic.git v5.5/dt64)

Change-Id: I450f6e3d9eee26b37703b55102f2e147c3f5182f