BACKPORT: arm64: dts: meson: g12: move audio bus out of g12-common
The base address of the audio bus and pdm device are different
between the g12 and sm1 SoC families. Overwriting the reg property
only would leave with confusing node names on the sm1.

Move the audio related devices to the g12 dtsi. The appropriate nodes
will be created for the sm1 later on.

Change-Id: Ic4b404485e550e7af2f1f51f674e4672c7fd9b57
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>